feat: parse call graph to the graphlib json (#2)
* feat: parse call graph to the graphlib json

Generator outputs a call graph as a string on stdout.  Functionality for parsing this output to graphlib graph structure was added.

* feat: examine jars in classpath

For correct matching, jar info is needed as well as class and function
info. Jars in class path are peeked in and files with .class are taken
and mapped to jar names. This info is then merged with a call graph. 
Classnames are moved from path-like convention to standard Java FQ class names.