feat: print stacktrace in addition to error message
The subprocess calling the Java call graph generator was not adding the
stack trace and we were missing out on important information from the
call graph generator in our logging facilities. Due to our internal
logging tools trimming the size of our messages, we publish up to 5 frames
of the stack trace to our logs.