feat: support `gradlew` for gradle sub projects
For gradle sub projects the `gradlew` util is defined on the root folder
of the project.
Today we are trying to get the `gradlew` util only from the target path
of the gradle project and missing it.

This change takes the gradle util as a parameter from the gradle
plugin, which gives the benefit of having it also deal with windows.